Refactor multimodal encoding and network code
